All the modules VLGraphics, VLVG, VLVolume, VLMolecule are now always built.
Before the user was given the possibility to optionally build VL modules to
reduce the size of VL, which was a single monolithic library. Since now we
have separate libraries for each module the user is free to choose which one
to use and therefore maintaining such optional building system became for
the most part a useless complication.

Optional build remains for plugins since they increase the size of VLCore
(2D plugins) and of VLGraphics (3D plugins) and might rely on 3rd party
libraries (PNG, JPG, TIFF etc). In the future we might want to split each
plugin in its own library but the usefulness of this approach remains
somehow limited for the moment.